Create To-Do-List
I have created a To-do list app as my project for the Encryptix internship program. For this task, I created a list of tasks and displayed it on screen and gave 5 options:
1. Add task
2. Update task
3. Delete task
4. View task
5. Exit 
To implement this functionality, I applied if and elif and printed them on the screen.

You have to select the number of tasks you want to add.
After that, it will give you the above given 5 options. If you want to add, update, delete, or view your list, then select the option. Otherwise, select the exit option.